+++ robotframework-ls/src/robotframework_ls/impl/code_action.py
@@ -189,9 +189,74 @@ def _undefined_variable_code_action(
                 }
                 yield _wrap_edit_in_command(
                     completion_context,
-                    "Create ${%s} local variable" % name,
+                    "Create local variable",
                     change,
                 )
+
+                # We can also create a variable in the variables section.
+                yield _undefined_variable_code_action_create_in_variables_section(
+                    completion_context, undefined_variable_data
+                )
+
+                if ast_utils.is_keyword_section(current_section):
+                    # We can also create an argument.
+                    pass
+
+
+def _undefined_variable_code_action_create_in_variables_section(
+    completion_context: ICompletionContext,
+    undefined_variable_data: ICustomDiagnosticDataUndefinedVariableTypedDict,
+) -> CodeActionTypedDict:
+    from robotframework_ls.impl import ast_utils
+
+    var_template = "${%s}    $__LSP_CURSOR_LOCATION__$\n" % (
+        undefined_variable_data["name"],
+    )
+    # Find if there's a variables section already...
+    variable_section = ast_utils.find_variable_section(completion_context.get_ast())
+    text_edit: TextEditTypedDict
+    if variable_section is None:
+        # We need to create the variables section too
+        current_section = completion_context.get_ast_current_section()
+        if current_section is None:
+            use_line = 0
+        else:
+            use_line = current_section.lineno - 1
+
+        text_edit = {
+            "range": {
+                "start": {"line": use_line, "character": 0},
+                "end": {"line": use_line, "character": 0},
+            },
+            "newText": f"*** Variables ***\n{var_template}\n",
+        }
+
+    else:
+        # We add the variable after the end of the existing variables section
+        last_stmt = None
+        for stmt in reversed(variable_section.body):
+            if not isinstance_name(stmt, "EmptyLine"):
+                last_stmt = stmt
+                break
+
+        if last_stmt is not None:
+            use_line = last_stmt.end_lineno
+        else:
+            use_line = variable_section.header.end_lineno
+
+        text_edit = {
+            "range": {
+                "start": {"line": use_line, "character": 0},
+                "end": {"line": use_line, "character": 0},
+            },
+            "newText": var_template,
+        }
+
+    return _wrap_edit_in_command(
+        completion_context,
+        "Create variable in variables section",
+        text_edit,
+    )
 
 
 def _wrap_edit_in_command(